Project refactor WIP
