+++ ios/ClothesStore/Home/HomeController.swift
@@ -10,10 +10,9 @@ import React
 
 class HomeController: BaseViewController {
 
-    let reactNativeEmitter = HomeBridge()
-
     private lazy var rnRootView: RCTRootView = {
-        let jsCodeLocation = Bundle.main.url(forResource: "main", withExtension: "jsbundle")!
+        let jsCodeLocation = URL(string: "http://localhost:8081/index.bundle?platform=ios")!
+        //let jsCodeLocation = Bundle.main.url(forResource: "main", withExtension: "jsbundle")!
         return RCTRootView(
             bundleURL: jsCodeLocation,
             moduleName: "RNHome",
@@ -22,12 +21,11 @@ class HomeController: BaseViewController {
         )
     }()
 
-    private var homePresenter: DataSourcePresenter<Product>!
+    var homePresenter: HomePresenter!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        homePresenter = DataSourcePresenter(dataControllerDelegate: nil, cartUpdateDelegate: self)
-
+        homePresenter = HomePresenter(dataControllerDelegate: nil, cartUpdateDelegate: self)
         self.view = rnRootView
 
     }
@@ -35,52 +33,19 @@ class HomeController: BaseViewController {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         self.navigationController?.navigationBar.isTranslucent = false
-        self.tabBarController?.title = "Home"
+        self.tabBarController?.title = "RN-Home"
     }
 
-    func fetchData() {
-        guard
-            let url = Bundle.main.url(forResource: "RNHomeDummyData", withExtension: "json"),
-            let data = try? Data(contentsOf: url),
-            let products = try? JSONSerialization.jsonObject(with: data, options: []) as? [[String: Any]]
-        else {
-            reactNativeEmitter.onErrorOccured(reason: "Unable to decode data!")
-            return
-        }
-
-        reactNativeEmitter.updateWishList(ids: homePresenter.wishlistManager.getWishListIds())
-        reactNativeEmitter.onDataRetrieved(data: products)
-    }
-
-    func addToCart(productId: Int) {
-        homePresenter.addToCart(id: productId)
-    }
-
-    func addToWishList(productDict: [String: Any]) {
-        if let deserializedProduct = Product.fromDictionary(dict: productDict) {
-            homePresenter.addToWishList(product: deserializedProduct)
-            reactNativeEmitter.updateWishList(ids: homePresenter.wishlistManager.getWishListIds())
-        } else {
-            reactNativeEmitter.onErrorOccured(reason: "Unable to add product (Err -40)!")
-        }
-    }
-
-    func removeFromWishList(productDict: [String: Any]) {
-        if let deserializedProduct = Product.fromDictionary(dict: productDict) {
-            homePresenter.removeFromWishList(product: deserializedProduct)
-            reactNativeEmitter.updateWishList(ids: homePresenter.wishlistManager.getWishListIds())
-        } else {
-            reactNativeEmitter.onErrorOccured(reason: "Unable to remove product (Err -41)!")
-        }
-    }
 }
+
+//Leaving this extension is deliberate, to show we can still access our view and mix UI+Server updates code.
 
 extension HomeController: CartUpdateDelegate {
     func onCartUpdateSuccess(message: String) {
-        reactNativeEmitter.onSuccessComplete(message: message)
+        homePresenter.reactNativeEmitter.onSuccessComplete(message: message)
     }
 
     func onCartUpdateFailed(reason: String) {
-        reactNativeEmitter.onErrorOccured(reason: reason)
+        homePresenter.reactNativeEmitter.onErrorOccured(reason: reason)
     }
 }
